+++ jsonjsdb/test/db/gdf9898fds/doc.json.js
@@ -1,23 +1,22 @@
-jsonjs.data['doc'] = 
-[
+jsonjs.data['doc'] = [
   {
-    "id": 1,
-    "name": "doc 1",
+    id: 1,
+    name: 'doc 1',
   },
   {
-    "id": 2,
-    "name": "doc 2",
+    id: 2,
+    name: 'doc 2',
   },
   {
-    "id": 3,
-    "name": "doc 3",
+    id: 3,
+    name: 'doc 3',
   },
   {
-    "id": 4,
-    "name": "doc 4",
+    id: 4,
+    name: 'doc 4',
   },
   {
-    "id": 5,
-    "name": "doc 5",
-  }
+    id: 5,
+    name: 'doc 5',
+  },
 ]