+++ jsonjsdb/test/db/gdf9898fds/tag.json.js
@@ -1,23 +1,22 @@
-jsonjs.data['tag'] = 
-[
+jsonjs.data['tag'] = [
   {
-    "id": 1,
-    "name": "tag 1"
+    id: 1,
+    name: 'tag 1',
   },
   {
-    "id": 2,
-    "name": "tag 2"
+    id: 2,
+    name: 'tag 2',
   },
   {
-    "id": 3,
-    "name": "tag 3"
+    id: 3,
+    name: 'tag 3',
   },
   {
-    "id": 4,
-    "name": "tag 4"
+    id: 4,
+    name: 'tag 4',
   },
   {
-    "id": 5,
-    "name": "tag 5"
-  }
+    id: 5,
+    name: 'tag 5',
+  },
 ]